repair global feed
